+++ lib/ratistics/monkey.rb
@@ -1,8 +1,0 @@
-require 'ratistics'
-
-require 'ratistics/monkey/array'
-require 'ratistics/monkey/hash'
-
-module Monkey
-
-end