+++ spec/ratistics/monkey_spec.rb
@@ -1,8 +1,0 @@
-require 'spec_helper'
-
-module Ratistics
-
-  describe Monkey do
-
-  end
-end